Final Commit

+++ get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: ricmanue <ricmanue@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/21 10:23:43 by ricmanue          #+#    #+#             */
-/*   Updated: 2024/06/25 09:50:31 by ricmanue         ###   ########.fr       */
+/*   Updated: 2024/06/26 15:07:40 by ricmanue         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define GET_NEXT_LINE_H
 
 # ifndef BUFFER_SIZE
-# define BUFFER_SIZE ((size_t)5)
+#  define BUFFER_SIZE 5
 # endif
 
 # include <stdlib.h>
@@ -27,7 +27,7 @@
 int		ft_strlen(char *c);
 char	*get_next_line(int fd);
 char	*ft_strjoin(char *s1, char *s2);
-int 	ft_newline (char *str);
-void	*ft_cleanread (char *actual_line, char *buffer_content);
+int		ft_newline(char *str);
+void	*ft_cleanread(char *actual_line, char *buffer_content);
 
 #endif
